Advanced_Phase_Deliverables/Task_2_Transcript_Generator/requirements.txt
streamlit
opencv-python-headless
pytesseract
pandas
fpdf
langchain-google-genai
python-dotenv
Pillow